Factor out `config` and `kv` modules from lib.rs
This commit is pure refactoring
